feat: cart & user order

+++ src/main/java/com/example/WarehouseContracts/dto/viewmodels/order/OrderViewModel.java
@@ -1,11 +1,10 @@
 package com.example.WarehouseContracts.dto.viewmodels.order;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record OrderViewModel(
-        String number,
-        String status,
-        LocalDateTime date,
-        BigDecimal totalPrice
+    String number,
+    String status,
+    LocalDateTime date,
+    Float totalPrice
 ) {}
